backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic
python-dotenv
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart